- Added test utils containing function to create BinaryFile from local file - Refactored tests to use new util
