+++ libs/extensions/std/exec/test/index.ts
@@ -3,3 +3,4 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 export * from './mocks';
+export * from './utils';
